# payment-service/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  version: 1.0.0

paths:
  /api/payments/getBalance/{userId}:
    get:
      tags:
        - payments
      summary: Получить баланс пользователя
      operationId: apiPaymentsGetBalanceUserIdGet
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Баланс пользователя"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '404':
          description: "Пользователь не найден"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: "Нет доступа"

  /api/payments/pay:
    post:
      summary: Оформить платеж
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'

  /api/payments/initBalance:
    post:
      summary: Инициализировать баланс
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitBalanceRequest'
      responses:
        '200':
          description: Баланс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'

components:
  schemas:
    Balance:
      type: object
      properties:
        userId:
          type: string
        balance:
          type: number
          format: double
      required:
        - userId
        - balance

    PaymentRequest:
      type: object
      properties:
        userId:
          type: string
        orderId:
          type: string
        amount:
          type: number
          format: double
      required:
        - userId
        - amount

    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        userId:
          type: string
        remainingBalance:
          type: number
          format: double
        orderId:
          type: string
      required:
        - success
        - userId

    InitBalanceRequest:
      type: object
      properties:
        userId:
          type: string
        initialBalance:
          type: number
          format: double
      required:
        - userId

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT